Fixed python::version to work with hiera data
python::version now uses hiera data for environment data when building versions with pyenv. CC and PYENV_ROOT are the only "default" environment data, but they could be overwritten in hiera if you wanted.

Added the ability to alias versions together. The class to do that is coming soon.

Found that pyenv was returing an error if the version had already been installed. Added the --skip-existing flag to pyenv to fix this issue.